Use html5 'required' attribute in bookmark form
There's some duplication between the required attribute in the template
and the validation defined in the Bookmark model.  I'm happy to live
with this for now.

Adding the required attribute causes Chrome to display a friendly little
'Please fill out this field' notification if you try to submit the form
with missing data.