Make Bookmarks responsible for creating tags
I did play around with a custom implementation of `Bookmark#tag_names`
but in the end it was easier just to store it in the database, and
accept the duplication.  This implementation makes for a simpler
bookmarks controller and I think it'll make it easier to add edit/update
functionality.