Make it easier to amend an existing bookmark
Using the bookmarklet to bookmark a page that you've previously
bookmarked will now result in you seeing the form to edit the existing
bookmark, rather than the form to create a new one.  This'll make it
more obvious that you've already seen this page.

*NOTE* This obviously won't work in situations where the content is
available at multiple URLs (e.g. one URL with a file extension and one
without).