Use the bookmarked url as the bookmark identifier
This should make it easy to construct a request to check for the existence of
a URL in javascript, e.g. `'/bookmarks/' +
encodeURIComponent(document.location.href)`.

The controller is a bit messy after this commit as I wanted to continue
supporting the viewing of bookmarks by id.

I'm not exactly sold on this URL structure as it feels a bit clumsy/messy but
it'll do for now.