Update gems
I ran `bundle update` to get the latest version of all the bundled gems.

I had to fix up a couple of tests as the new Capybara behaviour is to raise an
Ambiguous match error if multiple links are found with the same text. I
followed the advice of using `first(:link, 'text').click` (seen in the upgrade
guide[1]) for one of the tests, and changed the template to make the link text
more specific for the second.

[1] Edit the bookmark for http://foo.com